added import functionality on /importDB endpoint
