+++ controllers/settings_controller.go
@@ -84,3 +84,13 @@ func ExportDbToUsb(c *gin.Context) {
 	}
 	c.JSON(http.StatusOK, true)
 }
+
+func ImportDbFromUsb(c *gin.Context) {
+	_, err := services.ImportDbFromUsb("/home/dfxp/Desktop/AutomatedAccessControl/Repos/USF.DFX.ASM.API/")
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusOK, true)
+}
